- refactor model name and relocated non-model items
