+++ utils/estimate_job_count/care_homes.py
@@ -1,9 +1,9 @@
 from pyspark.ml.regression import GBTRegressionModel
 from utils.estimate_job_count.column_names import ESTIMATE_JOB_COUNT
-from utils.estimate_job_count.models.insert_predictions_into_locations import (
+from utils.estimate_job_count.insert_predictions_into_locations import (
     insert_predictions_into_locations,
 )
-from utils.estimate_job_count.models.r2_metric import generate_r2_metric
+from utils.estimate_job_count.r2_metric import generate_r2_metric
 from utils.prepare_locations_utils.job_calculator.job_calculator import (
     update_dataframe_with_identifying_rule,
 )
